don't remove .pyc files of targets
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@9349 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ site_scons/site_tools/autoclean_dir.py
@@ -27,7 +27,8 @@ def autoclean_dir(env, dirname):
     for (dirname, dirs, files) in os.walk(str(directory)):
         for filename in files:
             fullname = os.path.join(dirname, filename)
-            if not fullname in targets:
+            if not fullname in targets and not (fullname.endswith('.pyc')
+                    and fullname[:-1] in targets):
                 print env['CLEANUPCOMSTR'] % fullname
                 os.remove(fullname)
